requirements.txt
Flask~=1.1.2
python-dotenv~=0.15.0
SQLAlchemy~=1.4.1
WTForms~=2.3.3
marshmallow~=3.11.1
Werkzeug~=1.0.1